Refactor completion code into its own package
