+++ staging/src/k8s.io/kubectl/pkg/cmd/apiresources/apiresources.go
@@ -17,11 +17,8 @@ limitations under the License.
 package apiresources
 
 import (
-	"bytes"
 	"fmt"
 	"io"
-	"io/ioutil"
-	"os"
 	"sort"
 	"strings"
 
@@ -277,39 +274,4 @@ func (s sortableResource) compareValues(i, j int) (string, string) {
 		return s.resources[i].APIResource.Kind, s.resources[j].APIResource.Kind
 	}
 	return s.resources[i].APIGroup, s.resources[j].APIGroup
-}
-
-// CompGetResourceList returns the list of api resources which begin with `toComplete`.
-func CompGetResourceList(f cmdutil.Factory, cmd *cobra.Command, toComplete string) []string {
-	buf := new(bytes.Buffer)
-	streams := genericclioptions.IOStreams{In: os.Stdin, Out: buf, ErrOut: ioutil.Discard}
-	o := NewAPIResourceOptions(streams)
-
-	// Get the list of resources
-	o.Output = "name"
-	o.Cached = true
-	o.Verbs = []string{"get"}
-	// TODO:Should set --request-timeout=5s
-
-	// Ignore errors as the output may still be valid
-	o.RunAPIResources(cmd, f)
-
-	// Resources can be a comma-separated list.  The last element is then
-	// the one we should complete.  For example if toComplete=="pods,secre"
-	// we should return "pods,secrets"
-	prefix := ""
-	suffix := toComplete
-	lastIdx := strings.LastIndex(toComplete, ",")
-	if lastIdx != -1 {
-		prefix = toComplete[0 : lastIdx+1]
-		suffix = toComplete[lastIdx+1:]
-	}
-	var comps []string
-	resources := strings.Split(buf.String(), "\n")
-	for _, res := range resources {
-		if res != "" && strings.HasPrefix(res, suffix) {
-			comps = append(comps, fmt.Sprintf("%s%s", prefix, res))
-		}
-	}
-	return comps
 }
