+++ staging/src/k8s.io/kubectl/pkg/cmd/config/delete_context.go
@@ -23,7 +23,7 @@ import (
 	"github.com/spf13/cobra"
 	"k8s.io/client-go/tools/clientcmd"
 	cmdutil "k8s.io/kubectl/pkg/cmd/util"
-	"k8s.io/kubectl/pkg/util"
+	"k8s.io/kubectl/pkg/util/completion"
 	"k8s.io/kubectl/pkg/util/i18n"
 	"k8s.io/kubectl/pkg/util/templates"
 )
@@ -42,7 +42,7 @@ func NewCmdConfigDeleteContext(out, errOut io.Writer, configAccess clientcmd.Con
 		Short:                 i18n.T("Delete the specified context from the kubeconfig"),
 		Long:                  i18n.T("Delete the specified context from the kubeconfig."),
 		Example:               deleteContextExample,
-		ValidArgsFunction:     util.ContextCompletionFunc,
+		ValidArgsFunction:     completion.ContextCompletionFunc,
 		Run: func(cmd *cobra.Command, args []string) {
 			cmdutil.CheckErr(runDeleteContext(out, errOut, configAccess, cmd))
 		},
