+++ staging/src/k8s.io/kubectl/pkg/cmd/get/get.go
@@ -17,14 +17,11 @@ limitations under the License.
 package get
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net/url"
-	"os"
 	"strings"
 
 	"github.com/spf13/cobra"
@@ -45,7 +42,6 @@ import (
 	kubernetesscheme "k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
 	watchtools "k8s.io/client-go/tools/watch"
-	"k8s.io/kubectl/pkg/cmd/apiresources"
 	cmdutil "k8s.io/kubectl/pkg/cmd/util"
 	"k8s.io/kubectl/pkg/rawhttp"
 	"k8s.io/kubectl/pkg/scheme"
@@ -163,7 +159,7 @@ func NewCmdGet(parent string, f cmdutil.Factory, streams genericclioptions.IOStr
 		Short:                 i18n.T("Display one or many resources"),
 		Long:                  getLong + "\n\n" + cmdutil.SuggestAPIResources(parent),
 		Example:               getExample,
-		ValidArgsFunction:     ResourceTypeAndNameCompletionFunc(f, nil, true),
+		// ValidArgsFunction is set when this function is called so that we have access to the util package
 		Run: func(cmd *cobra.Command, args []string) {
 			cmdutil.CheckErr(o.Complete(f, cmd, args))
 			cmdutil.CheckErr(o.Validate(cmd))
@@ -850,141 +846,4 @@ func multipleGVKsRequested(infos []*resource.Info) bool {
 		}
 	}
 	return false
-}
-
-// CompGetResource gets the list of the resource specified which begin with `toComplete`.
-func CompGetResource(f cmdutil.Factory, cmd *cobra.Command, resourceName string, toComplete string) []string {
-	template := "{{ range .items  }}{{ .metadata.name }} {{ end }}"
-	return CompGetFromTemplate(&template, f, "", cmd, []string{resourceName}, toComplete)
-}
-
-// CompGetContainers gets the list of containers of the specified pod which begin with `toComplete`.
-func CompGetContainers(f cmdutil.Factory, cmd *cobra.Command, podName string, toComplete string) []string {
-	template := "{{ range .spec.initContainers }}{{ .name }} {{end}}{{ range .spec.containers  }}{{ .name }} {{ end }}"
-	return CompGetFromTemplate(&template, f, "", cmd, []string{"pod", podName}, toComplete)
-}
-
-// CompGetFromTemplate executes a Get operation using the specified template and args and returns the results
-// which begin with `toComplete`.
-func CompGetFromTemplate(template *string, f cmdutil.Factory, namespace string, cmd *cobra.Command, args []string, toComplete string) []string {
-	buf := new(bytes.Buffer)
-	streams := genericclioptions.IOStreams{In: os.Stdin, Out: buf, ErrOut: ioutil.Discard}
-	o := NewGetOptions("kubectl", streams)
-
-	// Get the list of names of the specified resource
-	o.PrintFlags.TemplateFlags.GoTemplatePrintFlags.TemplateArgument = template
-	format := "go-template"
-	o.PrintFlags.OutputFormat = &format
-
-	// Do the steps Complete() would have done.
-	// We cannot actually call Complete() or Validate() as these function check for
-	// the presence of flags, which, in our case won't be there
-	if namespace != "" {
-		o.Namespace = namespace
-		o.ExplicitNamespace = true
-	} else {
-		var err error
-		o.Namespace, o.ExplicitNamespace, err = f.ToRawKubeConfigLoader().Namespace()
-		if err != nil {
-			return nil
-		}
-	}
-
-	o.ToPrinter = func(mapping *meta.RESTMapping, outputObjects *bool, withNamespace bool, withKind bool) (printers.ResourcePrinterFunc, error) {
-		printer, err := o.PrintFlags.ToPrinter()
-		if err != nil {
-			return nil, err
-		}
-		return printer.PrintObj, nil
-	}
-
-	o.Run(f, cmd, args)
-
-	var comps []string
-	resources := strings.Split(buf.String(), " ")
-	for _, res := range resources {
-		if res != "" && strings.HasPrefix(res, toComplete) {
-			comps = append(comps, res)
-		}
-	}
-	return comps
-}
-
-// ResourceTypeAndNameCompletionFunc Returns a completion function that completes resource types
-// and resource names that match the toComplete prefix.  It supports the <type>/<name> form.
-func ResourceTypeAndNameCompletionFunc(f cmdutil.Factory, allowedTypes []string, allowRepeat bool) func(*cobra.Command, []string, string) ([]string, cobra.ShellCompDirective) {
-	return func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-		var comps []string
-		directive := cobra.ShellCompDirectiveNoFileComp
-
-		if len(args) > 0 && !strings.Contains(args[0], "/") {
-			// The first argument is of the form <type> (e.g., pods)
-			// All following arguments should be a resource name.
-			if allowRepeat || len(args) == 1 {
-				comps = CompGetResource(f, cmd, args[0], toComplete)
-
-				// Remove choices already on the command-line
-				if len(args) > 1 {
-					comps = cmdutil.Difference(comps, args[1:])
-				}
-			}
-		} else {
-			slashIdx := strings.Index(toComplete, "/")
-			if slashIdx == -1 {
-				if len(args) == 0 {
-					// We are completing the first argument.  We default to the normal
-					// <type> form (not the form <type>/<name>).
-					// So we suggest resource types and let the shell add a space after
-					// the completion.
-					if len(allowedTypes) == 0 {
-						comps = apiresources.CompGetResourceList(f, cmd, toComplete)
-					} else {
-						for _, c := range allowedTypes {
-							if strings.HasPrefix(c, toComplete) {
-								comps = append(comps, c)
-							}
-						}
-					}
-				} else {
-					// Here we know the first argument contains a / (<type>/<name>).
-					// All other arguments must also use that form.
-					if allowRepeat {
-						// Since toComplete does not already contain a / we know we are completing a
-						// resource type. Disable adding a space after the completion, and add the /
-						directive |= cobra.ShellCompDirectiveNoSpace
-
-						if len(allowedTypes) == 0 {
-							typeComps := apiresources.CompGetResourceList(f, cmd, toComplete)
-							for _, c := range typeComps {
-								comps = append(comps, fmt.Sprintf("%s/", c))
-							}
-						} else {
-							for _, c := range allowedTypes {
-								if strings.HasPrefix(c, toComplete) {
-									comps = append(comps, fmt.Sprintf("%s/", c))
-								}
-							}
-						}
-					}
-				}
-			} else {
-				// We are completing an argument of the form <type>/<name>
-				// and since the / is already present, we are completing the resource name.
-				if allowRepeat || len(args) == 0 {
-					resourceType := toComplete[:slashIdx]
-					toComplete = toComplete[slashIdx+1:]
-					nameComps := CompGetResource(f, cmd, resourceType, toComplete)
-					for _, c := range nameComps {
-						comps = append(comps, fmt.Sprintf("%s/%s", resourceType, c))
-					}
-
-					// Remove choices already on the command-line.
-					if len(args) > 0 {
-						comps = cmdutil.Difference(comps, args[0:])
-					}
-				}
-			}
-		}
-		return comps, directive
-	}
 }
