+++ staging/src/k8s.io/kubectl/pkg/util/completion/completion.go
@@ -14,14 +14,21 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package completion
 
 import (
+	"bytes"
 	"fmt"
+	"io/ioutil"
+	"os"
 	"strings"
 	"time"
 
 	"github.com/spf13/cobra"
+	"k8s.io/apimachinery/pkg/api/meta"
+	"k8s.io/cli-runtime/pkg/genericclioptions"
+	"k8s.io/cli-runtime/pkg/printers"
+	"k8s.io/kubectl/pkg/cmd/apiresources"
 	"k8s.io/kubectl/pkg/cmd/get"
 	cmdutil "k8s.io/kubectl/pkg/cmd/util"
 	"k8s.io/kubectl/pkg/polymorphichelpers"
@@ -30,7 +37,7 @@ import (
 
 var factory cmdutil.Factory
 
-// SetFactoryForCompletion Store the factory which is needed by the completion functions
+// SetFactoryForCompletion Store the factory which is needed by the completion functions.
 // Not all commands have access to the factory, so cannot pass it to the completion functions.
 func SetFactoryForCompletion(f cmdutil.Factory) {
 	factory = f
@@ -39,22 +46,21 @@ func SetFactoryForCompletion(f cmdutil.Factory) {
 // ResourceTypeAndNameCompletionFunc Returns a completion function that completes resource types
 // and resource names that match the toComplete prefix.  It supports the <type>/<name> form.
 func ResourceTypeAndNameCompletionFunc(f cmdutil.Factory) func(*cobra.Command, []string, string) ([]string, cobra.ShellCompDirective) {
-	// The logic is in the 'get' package because the 'get' command must be able to call it also
-	return get.ResourceTypeAndNameCompletionFunc(f, nil, true)
+	return resourceTypeAndNameCompletionFunc(f, nil, true)
 }
 
 // SpecifiedResourceTypeAndNameCompletionFunc Returns a completion function that completes resource
 // types limited to the specified allowedTypes, and resource names that match the toComplete prefix.
 // It allows for multiple resources. It supports the <type>/<name> form.
 func SpecifiedResourceTypeAndNameCompletionFunc(f cmdutil.Factory, allowedTypes []string) func(*cobra.Command, []string, string) ([]string, cobra.ShellCompDirective) {
-	return get.ResourceTypeAndNameCompletionFunc(f, allowedTypes, true)
+	return resourceTypeAndNameCompletionFunc(f, allowedTypes, true)
 }
 
 // SpecifiedResourceTypeAndNameNoRepeatCompletionFunc Returns a completion function that completes resource
 // types limited to the specified allowedTypes, and resource names that match the toComplete prefix.
 // It only allows for one resource. It supports the <type>/<name> form.
 func SpecifiedResourceTypeAndNameNoRepeatCompletionFunc(f cmdutil.Factory, allowedTypes []string) func(*cobra.Command, []string, string) ([]string, cobra.ShellCompDirective) {
-	return get.ResourceTypeAndNameCompletionFunc(f, allowedTypes, false)
+	return resourceTypeAndNameCompletionFunc(f, allowedTypes, false)
 }
 
 // ResourceNameCompletionFunc Returns a completion function that completes as a first argument
@@ -66,7 +72,7 @@ func ResourceNameCompletionFunc(f cmdutil.Factory, resourceType string) func(*co
 	return func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
 		var comps []string
 		if len(args) == 0 {
-			comps = get.CompGetResource(f, cmd, resourceType, toComplete)
+			comps = CompGetResource(f, cmd, resourceType, toComplete)
 		}
 		return comps, cobra.ShellCompDirectiveNoFileComp
 	}
@@ -98,7 +104,7 @@ func PodResourceNameAndContainerCompletionFunc(f cmdutil.Factory) func(*cobra.Co
 			comps, directive = doPodResourceCompletion(f, cmd, toComplete)
 		} else if len(args) == 1 {
 			podName := convertResourceNameToPodName(f, args[0])
-			comps = get.CompGetContainers(f, cmd, podName, toComplete)
+			comps = CompGetContainers(f, cmd, podName, toComplete)
 		}
 		return comps, directive
 	}
@@ -114,7 +120,7 @@ func ContainerCompletionFunc(f cmdutil.Factory) func(*cobra.Command, []string, s
 		// That first argument can also be of the form <type>/<name> so we need to convert it.
 		if len(args) > 0 {
 			podName := convertResourceNameToPodName(f, args[0])
-			comps = get.CompGetContainers(f, cmd, podName, toComplete)
+			comps = CompGetContainers(f, cmd, podName, toComplete)
 		}
 		return comps, cobra.ShellCompDirectiveNoFileComp
 	}
@@ -145,6 +151,64 @@ func UserCompletionFunc(cmd *cobra.Command, args []string, toComplete string) ([
 		return ListUsersInConfig(toComplete), cobra.ShellCompDirectiveNoFileComp
 	}
 	return nil, cobra.ShellCompDirectiveNoFileComp
+}
+
+// CompGetResource gets the list of the resource specified which begin with `toComplete`.
+func CompGetResource(f cmdutil.Factory, cmd *cobra.Command, resourceName string, toComplete string) []string {
+	template := "{{ range .items  }}{{ .metadata.name }} {{ end }}"
+	return CompGetFromTemplate(&template, f, "", cmd, []string{resourceName}, toComplete)
+}
+
+// CompGetContainers gets the list of containers of the specified pod which begin with `toComplete`.
+func CompGetContainers(f cmdutil.Factory, cmd *cobra.Command, podName string, toComplete string) []string {
+	template := "{{ range .spec.initContainers }}{{ .name }} {{end}}{{ range .spec.containers  }}{{ .name }} {{ end }}"
+	return CompGetFromTemplate(&template, f, "", cmd, []string{"pod", podName}, toComplete)
+}
+
+// CompGetFromTemplate executes a Get operation using the specified template and args and returns the results
+// which begin with `toComplete`.
+func CompGetFromTemplate(template *string, f cmdutil.Factory, namespace string, cmd *cobra.Command, args []string, toComplete string) []string {
+	buf := new(bytes.Buffer)
+	streams := genericclioptions.IOStreams{In: os.Stdin, Out: buf, ErrOut: ioutil.Discard}
+	o := get.NewGetOptions("kubectl", streams)
+
+	// Get the list of names of the specified resource
+	o.PrintFlags.TemplateFlags.GoTemplatePrintFlags.TemplateArgument = template
+	format := "go-template"
+	o.PrintFlags.OutputFormat = &format
+
+	// Do the steps Complete() would have done.
+	// We cannot actually call Complete() or Validate() as these function check for
+	// the presence of flags, which, in our case won't be there
+	if namespace != "" {
+		o.Namespace = namespace
+		o.ExplicitNamespace = true
+	} else {
+		var err error
+		o.Namespace, o.ExplicitNamespace, err = f.ToRawKubeConfigLoader().Namespace()
+		if err != nil {
+			return nil
+		}
+	}
+
+	o.ToPrinter = func(mapping *meta.RESTMapping, outputObjects *bool, withNamespace bool, withKind bool) (printers.ResourcePrinterFunc, error) {
+		printer, err := o.PrintFlags.ToPrinter()
+		if err != nil {
+			return nil, err
+		}
+		return printer.PrintObj, nil
+	}
+
+	o.Run(f, cmd, args)
+
+	var comps []string
+	resources := strings.Split(buf.String(), " ")
+	for _, res := range resources {
+		if res != "" && strings.HasPrefix(res, toComplete) {
+			comps = append(comps, res)
+		}
+	}
+	return comps
 }
 
 // ListContextsInConfig returns a list of context names which begin with `toComplete`
@@ -192,6 +256,120 @@ func ListUsersInConfig(toComplete string) []string {
 	return ret
 }
 
+// compGetResourceList returns the list of api resources which begin with `toComplete`.
+func compGetResourceList(f cmdutil.Factory, cmd *cobra.Command, toComplete string) []string {
+	buf := new(bytes.Buffer)
+	streams := genericclioptions.IOStreams{In: os.Stdin, Out: buf, ErrOut: ioutil.Discard}
+	o := apiresources.NewAPIResourceOptions(streams)
+
+	// Get the list of resources
+	o.Output = "name"
+	o.Cached = true
+	o.Verbs = []string{"get"}
+	// TODO:Should set --request-timeout=5s
+
+	// Ignore errors as the output may still be valid
+	o.RunAPIResources(cmd, f)
+
+	// Resources can be a comma-separated list.  The last element is then
+	// the one we should complete.  For example if toComplete=="pods,secre"
+	// we should return "pods,secrets"
+	prefix := ""
+	suffix := toComplete
+	lastIdx := strings.LastIndex(toComplete, ",")
+	if lastIdx != -1 {
+		prefix = toComplete[0 : lastIdx+1]
+		suffix = toComplete[lastIdx+1:]
+	}
+	var comps []string
+	resources := strings.Split(buf.String(), "\n")
+	for _, res := range resources {
+		if res != "" && strings.HasPrefix(res, suffix) {
+			comps = append(comps, fmt.Sprintf("%s%s", prefix, res))
+		}
+	}
+	return comps
+}
+
+// resourceTypeAndNameCompletionFunc Returns a completion function that completes resource types
+// and resource names that match the toComplete prefix.  It supports the <type>/<name> form.
+func resourceTypeAndNameCompletionFunc(f cmdutil.Factory, allowedTypes []string, allowRepeat bool) func(*cobra.Command, []string, string) ([]string, cobra.ShellCompDirective) {
+	return func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+		var comps []string
+		directive := cobra.ShellCompDirectiveNoFileComp
+
+		if len(args) > 0 && !strings.Contains(args[0], "/") {
+			// The first argument is of the form <type> (e.g., pods)
+			// All following arguments should be a resource name.
+			if allowRepeat || len(args) == 1 {
+				comps = CompGetResource(f, cmd, args[0], toComplete)
+
+				// Remove choices already on the command-line
+				if len(args) > 1 {
+					comps = cmdutil.Difference(comps, args[1:])
+				}
+			}
+		} else {
+			slashIdx := strings.Index(toComplete, "/")
+			if slashIdx == -1 {
+				if len(args) == 0 {
+					// We are completing the first argument.  We default to the normal
+					// <type> form (not the form <type>/<name>).
+					// So we suggest resource types and let the shell add a space after
+					// the completion.
+					if len(allowedTypes) == 0 {
+						comps = compGetResourceList(f, cmd, toComplete)
+					} else {
+						for _, c := range allowedTypes {
+							if strings.HasPrefix(c, toComplete) {
+								comps = append(comps, c)
+							}
+						}
+					}
+				} else {
+					// Here we know the first argument contains a / (<type>/<name>).
+					// All other arguments must also use that form.
+					if allowRepeat {
+						// Since toComplete does not already contain a / we know we are completing a
+						// resource type. Disable adding a space after the completion, and add the /
+						directive |= cobra.ShellCompDirectiveNoSpace
+
+						if len(allowedTypes) == 0 {
+							typeComps := compGetResourceList(f, cmd, toComplete)
+							for _, c := range typeComps {
+								comps = append(comps, fmt.Sprintf("%s/", c))
+							}
+						} else {
+							for _, c := range allowedTypes {
+								if strings.HasPrefix(c, toComplete) {
+									comps = append(comps, fmt.Sprintf("%s/", c))
+								}
+							}
+						}
+					}
+				}
+			} else {
+				// We are completing an argument of the form <type>/<name>
+				// and since the / is already present, we are completing the resource name.
+				if allowRepeat || len(args) == 0 {
+					resourceType := toComplete[:slashIdx]
+					toComplete = toComplete[slashIdx+1:]
+					nameComps := CompGetResource(f, cmd, resourceType, toComplete)
+					for _, c := range nameComps {
+						comps = append(comps, fmt.Sprintf("%s/%s", resourceType, c))
+					}
+
+					// Remove choices already on the command-line.
+					if len(args) > 0 {
+						comps = cmdutil.Difference(comps, args[0:])
+					}
+				}
+			}
+		}
+		return comps, directive
+	}
+}
+
 // doPodResourceCompletion Returns completions of:
 // 1- pod names that match the toComplete prefix
 // 2- resource types containing pods which match the toComplete prefix
@@ -201,7 +379,7 @@ func doPodResourceCompletion(f cmdutil.Factory, cmd *cobra.Command, toComplete s
 	slashIdx := strings.Index(toComplete, "/")
 	if slashIdx == -1 {
 		// Standard case, complete pod names
-		comps = get.CompGetResource(f, cmd, "pod", toComplete)
+		comps = CompGetResource(f, cmd, "pod", toComplete)
 
 		// Also include resource choices for the <type>/<name> form,
 		// but only for resources that contain pods
@@ -230,7 +408,7 @@ func doPodResourceCompletion(f cmdutil.Factory, cmd *cobra.Command, toComplete s
 		// Dealing with the <type>/<name> form, use the specified resource type
 		resourceType := toComplete[:slashIdx]
 		toComplete = toComplete[slashIdx+1:]
-		nameComps := get.CompGetResource(f, cmd, resourceType, toComplete)
+		nameComps := CompGetResource(f, cmd, resourceType, toComplete)
 		for _, c := range nameComps {
 			comps = append(comps, fmt.Sprintf("%s/%s", resourceType, c))
 		}
