+++ staging/src/k8s.io/kubectl/pkg/util/completion/completion_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package completion
 
 import (
 	"net/http"
